requirements.txt
backoff==1.10.0
beautifulsoup4==4.9.1
bs4==0.0.1==0.0.1
cachetools==4.1.0
chardet==3.0.4
dsub==0.3.8
google-api-core==1.17.0
google-api-python-client==1.8.3
google-auth==1.15.0
google-auth-httplib2==0.0.3
google-cloud-bigquery==1.25.0
google-cloud-core==1.3.0
google-cloud-storage==1.28.1
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
httplib2==0.17.3
idna==2.9
mock==3.0.5
parameterized==0.7.4
pip==20.1==20.1.1
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycurl==7.43.0.5
pydicom==2.0.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
retrying==1.3.3
rsa==4.0
setuptools==46.1.3
six==1.14.0
soupsieve==2.0.1
tabulate==0.8.7
uritemplate==3.0.1
urllib3==1.25.9
wheel==0.34.2